+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "auditore-forum",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -47,6 +47,7 @@
     "react-dom": "^19.2.3",
     "react-hook-form": "^7.69.0",
     "react-icons": "^5.5.0",
+    "sanitize-html": "^2.17.0",
     "sonner": "^2.0.7",
     "tailwind-merge": "^3.4.0",
     "zod": "^4.2.1"
@@ -57,6 +58,7 @@
     "@types/node": "^20.19.27",
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
+    "@types/sanitize-html": "^2.16.0",
     "eslint": "^9.39.2",
     "eslint-config-next": "15.3.5",
     "tailwindcss": "^4.1.18",
