+++ src/components/CreateTopicView.tsx
@@ -10,7 +10,6 @@ export function CreateTopicView({ category }: { category: string }) {
     content,
     setContent,
     isSubmitting,
-    error,
     handleTopicSubmit,
     addImage,
   } = useCreateTopic(category);
@@ -19,7 +18,6 @@ export function CreateTopicView({ category }: { category: string }) {
     <div className="flex w-full max-w-6xl mx-auto gap-6 p-4 text-white">
       <div className="flex-1">
         <h1 className="text-3xl font-bold mb-4">Criar Novo Tópico</h1>
-        {error && <p className="text-red-500 mb-4">{error}</p>}
         <PublishForm
           type="topic"
           onSubmit={() => handleTopicSubmit(true)}
