+++ src/components/index/OnlineUsers.tsx
@@ -14,7 +14,6 @@ export function OnlineUsers() {
   const auth = useAuth();
   const currentUser = auth?.user || null;
 
-  //  que seu hook retorna 'any' ou 'unknown' em onlineUsers
   const { onlineUsers, isConnected } = useOnlineUsers();
 
   const users: OnlineUser[] = Array.isArray(onlineUsers)
