+++ src/components/profile/FollowerListModal.tsx
@@ -1,16 +1,16 @@
 "use client";
 
 import Link from "next/link";
-import { useState } from "react";
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Loader2, X, AlertTriangle } from "lucide-react";
-import { UserPreview } from "@/schema/user";
 import { Button } from "../ui/button";
-import { useRemoveFollower } from "@/hooks/useRemoveFollower";
-import { toast } from "sonner";
 import { getRoleColor } from "@/utils/colors";
+import { useRemoveFollower } from "@/hooks/useRemoveFollower";
+import {
+  useFollowModalData,
+  useUnfollowMutation,
+} from "@/hooks/useFollowActions";
 
 interface FollowListModalProps {
   username: string;
@@ -19,81 +19,20 @@ interface FollowListModalProps {
   isOwnProfile: boolean;
 }
 
-const fetchModalUsers = async (
-  username: string,
-  listType: "followers" | "following"
-) => {
-  try {
-    const res = await fetch(`/api/follow/${username}/${listType}`, {
-      credentials: "include",
-    });
-    if (!res.ok) {
-      const errorData = await res.json().catch(() => ({}));
-      throw new Error(errorData.message || "Falha ao carregar a lista.");
-    }
-    const data = await res.json();
-    return data;
-  } catch (error: unknown) {
-    if (error instanceof Error) {
-      throw new Error(error.message);
-    }
-    throw new Error("Não foi possível carregar a lista de usuários.");
-  }
-};
-
 export function FollowListModal({
   username,
   listType,
   onClose,
   isOwnProfile,
 }: FollowListModalProps) {
-  const queryClient = useQueryClient();
-  const { removeFollower, isPending: isRemoving } = useRemoveFollower();
-  const [mutatingUsername, setMutatingUsername] = useState<string | null>(null);
-
-  const handleRemoveClick = (followerUsername: string) => {
-    removeFollower({
-      followerUsername,
-      profileOwnerUsername: username,
-    });
-  };
-
-  const { mutate: unfollowUser } = useMutation({
-    mutationFn: async (usernameToUnfollow: string) => {
-      setMutatingUsername(usernameToUnfollow);
-      const res = await fetch(`/api/follow/${usernameToUnfollow}/follow`, {
-        method: "DELETE",
-        credentials: "include",
-      });
-      if (!res.ok) {
-        const errorData = await res.json().catch(() => ({}));
-        throw new Error(errorData.message || "Falha ao deixar de seguir o usuário.");
-      }
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["followModalList"] });
-    },
-    onError: (error) => {
-      let errorMessage = "Ocorreu um erro ao deixar de seguir o usuário.";
-      if (error instanceof Error) {
-        errorMessage = error.message;
-      }
-      toast.error(errorMessage);
-    },
-    onSettled: () => {
-      setMutatingUsername(null);
-    },
-  });
-
   const {
     data: users,
     isLoading,
     error,
-  } = useQuery<UserPreview[], Error>({
-    queryKey: ["followModalList", username, listType],
-    queryFn: () => fetchModalUsers(username, listType),
-    enabled: !!username,
-  });
+  } = useFollowModalData(username, listType);
+  const { mutate: unfollow, variables: unfollowingUsername } =
+    useUnfollowMutation();
+  const { removeFollower, isPending: isRemoving } = useRemoveFollower();
 
   return (
     <div className="fixed inset-0 bg-black/50 backdrop-blur-sm z-50 flex justify-center items-center p-4">
@@ -104,56 +43,62 @@ export function FollowListModal({
           </CardTitle>
           <button
             onClick={onClose}
-            className="absolute top-3 right-3 text-gray-400 hover:text-red-500 cursor-pointer transition-opacity"
+            className="absolute top-3 right-3 text-gray-400 hover:text-red-500 transition-opacity"
           >
             <X size={24} />
           </button>
         </CardHeader>
+
         <CardContent className="max-h-[60vh] overflow-y-auto">
           {isLoading ? (
-            <div className="flex justify-center items-center p-8 text-white">
+            <div className="flex justify-center p-8">
               <Loader2 className="h-8 w-8 animate-spin text-blue-500" />
             </div>
           ) : error ? (
-            <div className="flex flex-col items-center justify-center p-8 text-red-500">
+            <div className="flex flex-col items-center p-8 text-red-500">
               <AlertTriangle className="h-6 w-6 mb-2" />
               <p>{error.message}</p>
             </div>
-          ) : users && users.length > 0 ? (
+          ) : users?.length ? (
             <ul className="space-y-2">
               {users.map((user) => (
                 <li
                   key={user.username}
-                  className="flex items-center justify-between gap-2 p-2 rounded-lg hover:bg-slate-700 transition-colors text-white"
+                  className="flex items-center justify-between gap-2 p-2 rounded-lg hover:bg-slate-700 text-white"
                 >
                   <Link
                     href={`/profile/${user.username}`}
                     onClick={onClose}
                     className="flex items-center gap-4 flex-1 min-w-0"
                   >
-                    <Avatar className="bg-slate-600">
-                      <AvatarImage
-                        src={user.avatar_url || undefined}
-                        alt={`avatar de ${user.username}`}
-                      />
+                    <Avatar>
+                      <AvatarImage src={user.avatar_url || undefined} />
                       <AvatarFallback className="bg-slate-600">
                         {user.username.charAt(0).toUpperCase()}
                       </AvatarFallback>
                     </Avatar>
-                    <span className="font-semibold truncate">
-                      <span className={getRoleColor(user.role)}>
-                        {user.username}
-                      </span>
+                    <span
+                      className={`font-semibold truncate ${getRoleColor(
+                        user.role
+                      )}`}
+                    >
+                      {user.username}
                     </span>
                   </Link>
-                  <div className="flex-shrink-0">
+
+                  <div className="shrink-0">
                     {isOwnProfile && listType === "followers" && (
                       <Button
                         variant="destructive"
                         size="sm"
-                        onClick={() => handleRemoveClick(user.username)}
+                        className="w-21.25"
+                        onClick={() =>
+                          removeFollower({
+                            followerUsername: user.username,
+                            profileOwnerUsername: username,
+                          })
+                        }
                         disabled={isRemoving}
-                        className="w-[80px] cursor-pointer"
                       >
                         {isRemoving ? (
                           <Loader2 className="h-4 w-4 animate-spin" />
@@ -162,15 +107,16 @@ export function FollowListModal({
                         )}
                       </Button>
                     )}
+
                     {isOwnProfile && listType === "following" && (
                       <Button
                         variant="destructive"
                         size="sm"
-                        onClick={() => unfollowUser(user.username)}
-                        disabled={mutatingUsername === user.username}
-                        className="w-[120px] cursor-poiner"
+                        className="w-30"
+                        onClick={() => unfollow(user.username)}
+                        disabled={unfollowingUsername === user.username}
                       >
-                        {mutatingUsername === user.username ? (
+                        {unfollowingUsername === user.username ? (
                           <Loader2 className="h-4 w-4 animate-spin" />
                         ) : (
                           "Deixar de seguir"
@@ -183,7 +129,7 @@ export function FollowListModal({
             </ul>
           ) : (
             <p className="text-center text-gray-400 py-4">
-              Nenhum usuário para mostrar.
+              Nenhum usuário encontrado.
             </p>
           )}
         </CardContent>
