+++ src/components/user-settings/DangerZoneCard.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState } from "react";
+import { useState, useEffectEvent } from "react";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
@@ -23,59 +23,29 @@ import {
   AlertDialogTrigger,
 } from "@/components/ui/alert-dialog";
 import { toast } from "sonner";
-import { handleApiError } from "@/utils/apiErrors";
+import { useDeleteAccount } from "@/hooks/useDeleteAccount";
 
 export function DangerZoneCard() {
   const [step, setStep] = useState<"initial" | "confirmPassword">("initial");
   const [password, setPassword] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
   const [isOpen, setIsOpen] = useState(false);
 
-  const handleOpenChange = (open: boolean) => {
+  const { mutate, isPending } = useDeleteAccount();
+
+  const handleOpenChange = useEffectEvent((open: boolean) => {
     setIsOpen(open);
     if (!open) {
-      setTimeout(() => {
-        setStep("initial");
-        setPassword("");
-      }, 200);
+      setStep("initial");
+      setPassword("");
     }
-  };
+  });
 
-  const handleDeleteAccount = async () => {
+  const handleConfirmDelete = () => {
     if (!password) {
       toast.error("Por favor, digite sua senha para confirmar.");
       return;
     }
-
-    setIsLoading(true);
-    try {
-      const res = await fetch(`/api/auth/delete-account`, {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ password }),
-        credentials: "include",
-      });
-      if (!res.ok) {
-        const errorData = await res.json().catch(() => ({}));
-        throw new Error("Falha ao deletar a conta.", {
-          cause: errorData.error,
-        });
-      }
-
-      toast.success("Conta deletada com sucesso. Você será redirecionado.");
-      setTimeout(() => {
-        window.location.href = "/";
-      }, 2000);
-    } catch (error) {
-      handleApiError(error, "Falha ao deletar a conta.");
-      toast.error(
-        "Falha ao deletar a conta. Verifique sua senha e tente novamente."
-      );
-    } finally {
-      setIsLoading(false);
-    }
+    mutate(password);
   };
 
   return (
@@ -100,6 +70,7 @@ export function DangerZoneCard() {
                 servidores.
               </AlertDialogDescription>
             </AlertDialogHeader>
+
             {step === "confirmPassword" && (
               <div className="my-4 space-y-2">
                 <Label htmlFor="password-confirm">
@@ -111,14 +82,19 @@ export function DangerZoneCard() {
                   placeholder="Sua senha..."
                   value={password}
                   onChange={(e) => setPassword(e.target.value)}
+                  disabled={isPending}
                   className="bg-slate-700 border border-slate-600 text-white"
                 />
               </div>
             )}
             <AlertDialogFooter>
-              <AlertDialogCancel className="bg-slate-700 border border-slate-600 hover:bg-slate-600 cursor-pointer hover:text-white">
+              <AlertDialogCancel
+                className="bg-slate-700 border border-slate-600 hover:bg-slate-600 cursor-pointer"
+                disabled={isPending}
+              >
                 Cancelar
               </AlertDialogCancel>
+
               {step === "initial" ? (
                 <AlertDialogAction
                   className="bg-destructive text-destructive-foreground hover:bg-destructive/90 cursor-pointer"
@@ -132,10 +108,10 @@ export function DangerZoneCard() {
               ) : (
                 <AlertDialogAction
                   className="bg-destructive text-destructive-foreground hover:bg-destructive/50"
-                  onClick={handleDeleteAccount}
-                  disabled={isLoading}
+                  onClick={handleConfirmDelete}
+                  disabled={isPending}
                 >
-                  {isLoading ? "Excluindo..." : "Confirmar Exclusão"}
+                  {isPending ? "Excluindo..." : "Confirmar Exclusão"}
                 </AlertDialogAction>
               )}
             </AlertDialogFooter>
