+++ src/hooks/useForumData.ts
@@ -1,13 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { DashboardData } from "@/schema/forum";
-
-const fetchForumDashboardData = async (): Promise<DashboardData> => {
-  const res = await fetch(`/api/forum/data`);
-  if (!res.ok) {
-    throw new Error("Erro ao carregar os dados do painel do fórum.");
-  }
-  return res.json();
-};
+import { fetchForumDashboardData } from "@/app/api/endpoints/category";
 
 export const useForumData = () => {
   return useQuery<DashboardData>({
