+++ src/hooks/useUserStats.ts
@@ -1,24 +1,11 @@
-"use client";
+import { useQuery } from "@tanstack/react-query";
+import { getUserStatsApi } from "@/app/api/endpoints/profiles";
 
-import { useQuery } from '@tanstack/react-query';
-import { getUserStats } from '@/services/profile';
-import { UserStats } from '@/schema/user';
-
-export function useUserStats(username?: string) {
-  const { 
-    data: stats,
-    isLoading, 
-    error,
-    isError,
-  } = useQuery<UserStats, Error>({
-    queryKey: ['userStats', username], 
-    queryFn: () => getUserStats(username!),
+export const useUserStats = (username: string) => {
+  return useQuery({
+    queryKey: ["userStats", username],
+    queryFn: () => getUserStatsApi(username),
     enabled: !!username,
+    staleTime: 1000 * 60 * 5,
   });
-
-  return { 
-    stats, 
-    isLoading, 
-    error: isError ? error.message : null 
-  };
-}
+};
