+++ src/schema/forum.ts
@@ -4,8 +4,8 @@ export const MemberSchema = z.object({
   id: z.string(),
   username: z.string(),
   role: z.string(),
-  avatar_url: z.string().nullish(),
-  joined_at: z.string(),
+  avatar_url: z.string().nullish().optional(),
+  joined_at: z.string().nullish(),
   last_login: z.string().nullish(),
   mensagens_count: z.number().optional(),
   messages: z.number(),
